COM-2530 move file + add default local + fix fromMiscToCompaniDate with utc
